Reduce size of cat by putting ERRNO_MIN in sp
I didn't want to use the stack pointer for popping items off the stack
because the AMD64 System V ABI requires a 16-byte alignment. Instead, I
used rbp for that. I used a constant ERRNO_MIN = -4096, which is a nice
multiple of 16. I think the kernel needs the stack pointer to be on the
stack, so I just changed the last 16 bytes of rsp to ERRNO_MIN.